# public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "imperia-api",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Mapped server"
        },
        {
            "url": "http://localhost:8080",
            "description": "Current server"
        },
        {
            "url": "https://imperia-api.herokuapp.com",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/api/comments": {
            "get": {
                "tags": [
                    "comments"
                ],
                "summary": "Index comments.",
                "operationId": "indexComments",
                "parameters": [
                    {
                        "name": "page[size]",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/PageSize"
                        }
                    },
                    {
                        "name": "page[number]",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/PageNumber"
                        }
                    },
                    {
                        "name": "filter[commentable_id]",
                        "in": "query",
                        "description": "Comment target morph lass id.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "filter[commentable_type]",
                        "in": "query",
                        "description": "Comment target class slug. Examples: `customers`, `banquets`,\n    `orders`, `product-field-orders`, `service-field-orders`, `space-field-orders`, `ticket-field-orders`",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "customers"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Index comments response object.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IndexCommentResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "comments"
                ],
                "summary": "Store comment.",
                "operationId": "storeComment",
                "requestBody": {
                    "description": "Store comment request object.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCommentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create comment response object.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StoreCommentResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/comments/{id}": {
            "get": {
                "tags": [
                    "comments"
                ],
                "summary": "Show comment by id.",
                "operationId": "showComment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the comment.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show commet response object.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShowCommentResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "comments"
                ],
                "summary": "Delete comment.",
                "operationId": "destroyComment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the comment.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete comment response object.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DestroyCommentResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "comments"
                ],
                "summary": "Update comment.",
                "operationId": "updateComment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the comment.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "Update comment request object.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCommentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Update comment response object.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateCommentResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UnauthenticatedResponse": {
                "description": "Unauthenticated response object.",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated"
                    }
                },
                "type": "object"
            },
...
            "IndexCommentResponse": {
                "description": "Index comments response object.",
                "required": [
                    "data",
                    "meta",
                    "message"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Comment"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    }
                },
                "type": "object"
            },
            "ShowCommentResponse": {
                "description": "Show comment response object.",
                "required": [
                    "data",
                    "message"
                ],
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Comment"
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    }
                },
                "type": "object"
            },
            "StoreCommentResponse": {
                "description": "Store comment response object.",
                "required": [
                    "data",
                    "message"
                ],
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Comment"
                    },
                    "message": {
                        "type": "string",
                        "example": "Created"
                    }
                },
                "type": "object"
            },
            "UpdateCommentResponse": {
                "description": "Update comment response object.",
                "required": [
                    "data",
                    "message"
                ],
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Comment"
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    }
                },
                "type": "object"
            },
            "DestroyCommentResponse": {
                "description": "Delete comment response object.",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Deleted"
                    }
                },
                "type": "object"
            },
...
            "PageSize": {
                "description": "Max number of results on page.",
                "type": "integer",
                "example": 25
            },
            "PageNumber": {
                "description": "Page number to be returned.",
                "type": "integer",
                "example": 1
            },
...
            "StoreCommentRequest": {
                "description": "Store comment request",
                "required": [
                    "text",
                    "commentable_id",
                    "commentable_type"
                ],
                "properties": {
                    "text": {
                        "type": "string",
                        "example": "This is a comment text..."
                    },
                    "commentable_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "commentable_type": {
                        "description": "Value for it must be obrained from the target resource's type property.",
                        "type": "string",
                        "example": "customers"
                    }
                },
                "type": "object"
            },
            "UpdateCommentRequest": {
                "description": "Update comment request",
                "required": [
                    "text"
                ],
                "properties": {
                    "text": {
                        "type": "string",
                        "example": "This is a comment text..."
                    }
                },
                "type": "object"
            },
...
            "Comment": {
                "description": "Comment resource object",
                "required": [
                    "id",
                    "type",
                    "text",
                    "commentable_id",
                    "commentable_type"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "example": "comments"
                    },
                    "text": {
                        "type": "string",
                        "example": "This is a comment."
                    },
                    "commentable_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "commentable_type": {
                        "type": "string",
                        "example": "products"
                    }
                },
                "type": "object"
            },
...
            "Links": {
                "description": "Pagination meta links object.",
                "required": [
                    "first",
                    "prev",
                    "self",
                    "next",
                    "last"
                ],
                "properties": {
                    "first": {
                        "type": "string"
                    },
                    "prev": {
                        "nullable": "true",
                        "type": "string",
                        "example": null
                    },
                    "self": {
                        "type": "string"
                    },
                    "next": {
                        "nullable": "true",
                        "type": "string",
                        "example": null
                    },
                    "last": {
                        "nullable": "true",
                        "type": "string",
                        "example": null
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "description": "Pagination meta object.",
                "required": [
                    "from",
                    "to",
                    "total",
                    "path",
                    "per_page",
                    "current_page",
                    "last_page",
                    "links"
                ],
                "properties": {
                    "from": {
                        "nullable": "true",
                        "type": "integer",
                        "example": 1
                    },
                    "to": {
                        "nullable": "true",
                        "type": "integer",
                        "example": 1
                    },
                    "total": {
                        "type": "integer",
                        "example": 30
                    },
                    "path": {
                        "nullable": "true",
                        "type": "integer",
                        "example": null
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 2
                    },
                    "links": {
                        "$ref": "#/components/schemas/Links"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}